set VERSION to GStreamer Core Version
The documentation is very unclear about this, implying that this 'version' field of the plugin definition macro where this variable would be used is actually different (independent) from the gstreamer version.  In fact, it is seemingly supposed to be the full
version string (of gstreamer core) that the plugin was compiled against, not the independent version of the plugin itself.